Introduce datapoint manager
